requirements.txt
openai
langchain
jinja2
markdown
python-frontmatter
pyyaml
schedule
gitpython
pytrends

